+++ src/strojure/undertow/server.clj
@@ -89,7 +89,7 @@
             {:host {\"app1.company.com\" (my-handler :app1-handler)
                     \"app2.company.com\" (my-handler :app2-handler)}})
           ;; Enable sessions for next handlers (above).
-          (handler/session-attachment {})
+          (handler/session {})
           ;; Path specific handlers.
           (handler/path {:prefix {\"static\" (handler/resource {:resource-manager :class-path
                                                               :prefix \"public/static\"})}
@@ -117,7 +117,7 @@
                            :prefix \"public/static\"}}
         :exact {\"websocket\" {:type handler/websocket :callback websocket-callback}}}
        ;; Enable sessions for next handlers.
-       {:type handler/session-attachment}
+       {:type handler/session}
        ;; The handlers for app hostnames.
        {:type handler/virtual-host
         :host {\"webapi.company.com\" (my-handler :webapi-handler)}}