+++ CHANGELOG.md
@@ -8,7 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 Release date `Unreleased`
 
-- 
+- Rename `handler/session-attachment` to `session`
 
 ## `1.0.47-beta.6`